refactor(mainnet): add pallet_preimage to utility
